+++ package.json
@@ -7,7 +7,9 @@
     "lint:js": "eslint .",
     "lint:css": "stylelint blocks/**/*.css styles/*.css",
     "lint": "npm run lint:js && npm run lint:css",
-    "build": "npm run build:css",
+    "build": "npm run build:css && npm run build:script:js && npm run build:franklin:js",
+    "build:franklin:js": "uglifyjs ./scripts/lib-franklin-dev.js -o ./scripts/lib-franklin.js -c -m",
+    "build:script:js": "uglifyjs ./scripts/scripts-dev.js -o ./scripts/scripts.js -c -m",
     "build:css": "npx tailwindcss ./styles/tailwind.css -o ./styles/styles.css --minify",
     "dev": "concurrently --kill-others \"npx tailwindcss ./styles/tailwind.css -o ./styles/styles.css --minify --watch\" \"aem up --print-index \"",
     "semantic-release": "semantic-release --debug",
