+++ scripts/lib-franklin.js
@@ -1,886 +1,1 @@
-/* eslint-disable max-classes-per-file */
-/*
- * Copyright 2022 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-
-/**
- * log RUM if part of the sample.
- * @param {string} checkpoint identifies the checkpoint in funnel
- * @param {Object} data additional data for RUM sample
- */
-export function sampleRUM(checkpoint, data = {}) {
-  sampleRUM.defer = sampleRUM.defer || [];
-  const defer = (fnname) => {
-    sampleRUM[fnname] = sampleRUM[fnname] || ((...args) => sampleRUM.defer.push({ fnname, args }));
-  };
-  sampleRUM.drain = sampleRUM.drain
-    || ((dfnname, fn) => {
-      sampleRUM[dfnname] = fn;
-      sampleRUM.defer
-        .filter(({ fnname }) => dfnname === fnname)
-        .forEach(({ fnname, args }) => sampleRUM[fnname](...args));
-    });
-  sampleRUM.always = sampleRUM.always || [];
-  sampleRUM.always.on = (chkpnt, fn) => {
-    sampleRUM.always[chkpnt] = fn;
-  };
-  sampleRUM.on = (chkpnt, fn) => {
-    sampleRUM.cases[chkpnt] = fn;
-  };
-  defer('observe');
-  defer('cwv');
-  try {
-    window.hlx = window.hlx || {};
-    if (!window.hlx.rum) {
-      const usp = new URLSearchParams(window.location.search);
-      const weight = usp.get('rum') === 'on' ? 1 : 20; // with parameter, weight is 1. Defaults to 100.
-      const id = Array.from({ length: 75 }, (_, i) => String.fromCharCode(48 + i))
-        .filter((a) => /\d|[A-Z]/i.test(a))
-        .filter(() => Math.random() * 75 > 70)
-        .join('');
-      const random = Math.random();
-      const isSelected = random * weight < 1;
-      const firstReadTime = Date.now();
-      const urlSanitizers = {
-        full: () => window.location.href,
-        origin: () => window.location.origin,
-        path: () => window.location.href.replace(/\?.*$/, ''),
-      };
-      // eslint-disable-next-line object-curly-newline, max-len
-      window.hlx.rum = {
-        weight,
-        id,
-        random,
-        isSelected,
-        firstReadTime,
-        sampleRUM,
-        sanitizeURL: urlSanitizers[window.hlx.RUM_MASK_URL || 'path'],
-      };
-    }
-    const { weight, id, firstReadTime } = window.hlx.rum;
-    if (window.hlx && window.hlx.rum && window.hlx.rum.isSelected) {
-      const knownProperties = [
-        'weight',
-        'id',
-        'referer',
-        'checkpoint',
-        't',
-        'source',
-        'target',
-        'cwv',
-        'CLS',
-        'FID',
-        'LCP',
-        'INP',
-      ];
-      const sendPing = (pdata = data) => {
-        // eslint-disable-next-line object-curly-newline, max-len, no-use-before-define
-        const body = JSON.stringify(
-          {
-            weight,
-            id,
-            referer: window.hlx.rum.sanitizeURL(),
-            checkpoint,
-            t: Date.now() - firstReadTime,
-            ...data,
-          },
-          knownProperties,
-        );
-        const url = `https://rum.hlx.page/.rum/${weight}`;
-        // eslint-disable-next-line no-unused-expressions
-        navigator.sendBeacon(url, body);
-        // eslint-disable-next-line no-console
-        console.debug(`ping:${checkpoint}`, pdata);
-      };
-      sampleRUM.cases = sampleRUM.cases || {
-        cwv: () => sampleRUM.cwv(data) || true,
-        lazy: () => {
-          // use classic script to avoid CORS issues
-          const script = document.createElement('script');
-          script.src = 'https://rum.hlx.page/.rum/@adobe/helix-rum-enhancer@^1/src/index.js';
-          document.head.appendChild(script);
-          return true;
-        },
-      };
-      sendPing(data);
-      if (sampleRUM.cases[checkpoint]) {
-        sampleRUM.cases[checkpoint]();
-      }
-    }
-    if (sampleRUM.always[checkpoint]) {
-      sampleRUM.always[checkpoint](data);
-    }
-  } catch (error) {
-    // something went wrong
-  }
-}
-
-/**
- * Loads a CSS file.
- * @param {string} href URL to the CSS file
- */
-export async function loadCSS(href) {
-  return new Promise((resolve, reject) => {
-    if (!document.querySelector(`head > link[href="${href}"]`)) {
-      const link = document.createElement('link');
-      link.rel = 'stylesheet';
-      link.href = href;
-      link.onload = resolve;
-      link.onerror = reject;
-      document.head.append(link);
-    } else {
-      resolve();
-    }
-  });
-}
-
-/**
- * Loads a non module JS file.
- * @param {string} src URL to the JS file
- * @param {Object} attrs additional optional attributes
- */
-
-export async function loadScript(src, attrs) {
-  return new Promise((resolve, reject) => {
-    if (!document.querySelector(`head > script[src="${src}"]`)) {
-      const script = document.createElement('script');
-      script.src = src;
-      if (attrs) {
-      // eslint-disable-next-line no-restricted-syntax, guard-for-in
-        for (const attr in attrs) {
-          script.setAttribute(attr, attrs[attr]);
-        }
-      }
-      script.onload = resolve;
-      script.onerror = reject;
-      document.head.append(script);
-    } else {
-      resolve();
-    }
-  });
-}
-
-/**
- * Retrieves the content of metadata tags.
- * @param {string} name The metadata name (or property)
- * @returns {string} The metadata value(s)
- */
-export function getMetadata(name) {
-  const attr = name && name.includes(':') ? 'property' : 'name';
-  const meta = [...document.head.querySelectorAll(`meta[${attr}="${name}"]`)].map((m) => m.content).join(', ');
-  return meta || '';
-}
-
-/**
- * Sanitizes a string for use as class name.
- * @param {string} name The unsanitized string
- * @returns {string} The class name
- */
-export function toClassName(name) {
-  return typeof name === 'string'
-    ? name.toLowerCase().replace(/[^0-9a-z]/gi, '-').replace(/-+/g, '-').replace(/^-|-$/g, '')
-    : '';
-}
-
-/**
- * Sanitizes a string for use as a js property name.
- * @param {string} name The unsanitized string
- * @returns {string} The camelCased name
- */
-export function toCamelCase(name) {
-  return toClassName(name).replace(/-([a-z])/g, (g) => g[1].toUpperCase());
-}
-
-/**
- * Gets all the metadata elements that are in the given scope.
- * @param {String} scope The scope/prefix for the metadata
- * @returns an array of HTMLElement nodes that match the given scope
- */
-export function getAllMetadata(scope) {
-  return [...document.head.querySelectorAll(`meta[property^="${scope}:"],meta[name^="${scope}-"]`)]
-    .reduce((res, meta) => {
-      const id = toClassName(meta.name
-        ? meta.name.substring(scope.length + 1)
-        : meta.getAttribute('property').split(':')[1]);
-      res[id] = meta.getAttribute('content');
-      return res;
-    }, {});
-}
-
-const ICONS_CACHE = {};
-/**
- * Replace icons with inline SVG and prefix with codeBasePath.
- * @param {Element} [element] Element containing icons
- */
-export async function decorateIcons(element) {
-  // Prepare the inline sprite
-  let svgSprite = document.getElementById('franklin-svg-sprite');
-  if (!svgSprite) {
-    const div = document.createElement('div');
-    div.innerHTML = '<svg xmlns="http://www.w3.org/2000/svg" id="franklin-svg-sprite" style="display: none"></svg>';
-    svgSprite = div.firstElementChild;
-    document.body.append(div.firstElementChild);
-  }
-
-  // Download all new icons
-  const icons = [...element.querySelectorAll('span.icon')];
-  await Promise.all(icons.map(async (span) => {
-    const iconName = Array.from(span.classList).find((c) => c.startsWith('icon-')).substring(5);
-    if (!ICONS_CACHE[iconName]) {
-      ICONS_CACHE[iconName] = true;
-      try {
-        let iconSource = `${window.hlx.codeBasePath}/icons/${iconName}.svg`;
-        if (iconName.startsWith('dam-')) {
-          const isPublicDomain = window.location.hostname.includes('lifesciences.danaher.com');
-          iconSource = isPublicDomain ? '' : 'https://lifesciences.danaher.com';
-          iconSource += `/content/dam/danaher/system/icons/${iconName.substring(4).replace('_', ' ')}.svg`;
-        }
-        const response = await fetch(iconSource);
-        if (!response.ok) {
-          ICONS_CACHE[iconName] = false;
-          return;
-        }
-        // Styled icons don't play nice with the sprite approach because of shadow dom isolation
-        const svg = await response.text();
-        if (svg.match(/(<style | class=)/)) {
-          ICONS_CACHE[iconName] = { styled: true, html: svg };
-        } else {
-          ICONS_CACHE[iconName] = {
-            html: svg
-              .replace('<svg', `<symbol id="icons-sprite-${iconName}"`)
-              .replace(/ width=".*?"/, '')
-              .replace(/ height=".*?"/, '')
-              .replace('</svg>', '</symbol>'),
-          };
-        }
-      } catch (error) {
-        ICONS_CACHE[iconName] = false;
-        // eslint-disable-next-line no-console
-        console.error(error);
-      }
-    }
-  }));
-
-  const symbols = Object
-    .keys(ICONS_CACHE).filter((k) => !svgSprite.querySelector(`#icons-sprite-${k}`))
-    .map((k) => ICONS_CACHE[k])
-    .filter((v) => !v.styled)
-    .map((v) => v.html)
-    .join('\n');
-  svgSprite.innerHTML += symbols;
-
-  icons.forEach((span) => {
-    const iconName = Array.from(span.classList).find((c) => c.startsWith('icon-')).substring(5);
-    const parent = span.firstElementChild?.tagName === 'A' ? span.firstElementChild : span;
-    // Styled icons need to be inlined as-is, while unstyled ones can leverage the sprite
-    if (ICONS_CACHE[iconName].styled) {
-      parent.innerHTML = ICONS_CACHE[iconName].html;
-    } else {
-      parent.innerHTML = `<svg xmlns="http://www.w3.org/2000/svg"><use href="#icons-sprite-${iconName}"/></svg>`;
-    }
-  });
-}
-
-/**
- * Gets placeholders object.
- * @param {string} [prefix] Location of placeholders
- * @returns {object} Window placeholders object
- */
-export async function fetchPlaceholders(prefix = 'default') {
-  window.placeholders = window.placeholders || {};
-  const loaded = window.placeholders[`${prefix}-loaded`];
-  if (!loaded) {
-    window.placeholders[`${prefix}-loaded`] = new Promise((resolve, reject) => {
-      fetch(`${prefix === 'default' ? '' : prefix}/placeholders.json`)
-        .then((resp) => {
-          if (resp.ok) {
-            return resp.json();
-          }
-          throw new Error(`${resp.status}: ${resp.statusText}`);
-        }).then((json) => {
-          const placeholders = {};
-          json.data
-            .filter((placeholder) => placeholder.Key)
-            .forEach((placeholder) => {
-              placeholders[toCamelCase(placeholder.Key)] = placeholder.Text;
-            });
-          window.placeholders[prefix] = placeholders;
-          resolve();
-        }).catch((error) => {
-          // error loading placeholders
-          window.placeholders[prefix] = {};
-          reject(error);
-        });
-    });
-  }
-  await window.placeholders[`${prefix}-loaded`];
-  return window.placeholders[prefix];
-}
-
-/**
- * Decorates a block.
- * @param {Element} block The block element
- */
-export function decorateBlock(block) {
-  const shortBlockName = block.classList[0];
-  if (shortBlockName) {
-    block.classList.add('block');
-    block.dataset.blockName = shortBlockName;
-    block.dataset.blockStatus = 'initialized';
-    const blockWrapper = block.parentElement;
-    blockWrapper.classList.add(`${shortBlockName}-wrapper`);
-    const section = block.closest('.section');
-    if (section) section.classList.add(`${shortBlockName}-container`);
-  }
-}
-
-/**
- * Extracts the config from a block.
- * @param {Element} block The block element
- * @returns {object} The block config
- */
-export function readBlockConfig(block) {
-  const config = {};
-  block.querySelectorAll(':scope > div').forEach((row) => {
-    if (row.children) {
-      const cols = [...row.children];
-      if (cols[1]) {
-        const col = cols[1];
-        const name = toClassName(cols[0].textContent);
-        let value = '';
-        if (col.querySelector('a')) {
-          const as = [...col.querySelectorAll('a')];
-          if (as.length === 1) {
-            value = as[0].href;
-          } else {
-            value = as.map((a) => a.href);
-          }
-        } else if (col.querySelector('img')) {
-          const imgs = [...col.querySelectorAll('img')];
-          if (imgs.length === 1) {
-            value = imgs[0].src;
-          } else {
-            value = imgs.map((img) => img.src);
-          }
-        } else if (col.querySelector('p')) {
-          const ps = [...col.querySelectorAll('p')];
-          if (ps.length === 1) {
-            value = ps[0].textContent;
-          } else {
-            value = ps.map((p) => p.textContent);
-          }
-        } else value = row.children[1].textContent;
-        config[name] = value;
-      }
-    }
-  });
-  return config;
-}
-
-/**
- * Decorates all sections in a container element.
- * @param {Element} main The container element
- */
-export function decorateSections(main) {
-  main.querySelectorAll(':scope > div').forEach((section) => {
-    const wrappers = [];
-    let defaultContent = false;
-    [...section.children].forEach((e) => {
-      if (e.tagName === 'DIV' || !defaultContent) {
-        const wrapper = document.createElement('div');
-        wrappers.push(wrapper);
-        defaultContent = e.tagName !== 'DIV';
-        if (defaultContent) wrapper.classList.add('default-content-wrapper');
-      }
-      wrappers[wrappers.length - 1].append(e);
-    });
-    wrappers.forEach((wrapper) => section.append(wrapper));
-    section.classList.add('section');
-    section.dataset.sectionStatus = 'initialized';
-    section.style.display = 'none';
-
-    /* process section metadata */
-    const sectionMeta = section.querySelector('div.section-metadata');
-    if (sectionMeta) {
-      const meta = readBlockConfig(sectionMeta);
-      Object.keys(meta).forEach((key) => {
-        if (key === 'style') {
-          const styles = meta.style.split(',').map((style) => toClassName(style.trim()));
-          styles.forEach((style) => section.classList.add(style));
-        } else {
-          section.dataset[toCamelCase(key)] = meta[key];
-        }
-      });
-      sectionMeta.parentNode.remove();
-    }
-  });
-}
-
-/**
- * Updates all section status in a container element.
- * @param {Element} main The container element
- */
-export function updateSectionsStatus(main) {
-  const sections = [...main.querySelectorAll(':scope > div.section')];
-  for (let i = 0; i < sections.length; i += 1) {
-    const section = sections[i];
-    const status = section.dataset.sectionStatus;
-    if (status !== 'loaded') {
-      const loadingBlock = section.querySelector('.block[data-block-status="initialized"], .block[data-block-status="loading"]');
-      if (loadingBlock) {
-        section.dataset.sectionStatus = 'loading';
-        break;
-      } else {
-        section.dataset.sectionStatus = 'loaded';
-        section.style.display = null;
-      }
-    }
-  }
-}
-
-/**
- * Decorates all blocks in a container element.
- * @param {Element} main The container element
- */
-export function decorateBlocks(main) {
-  main
-    .querySelectorAll('div.section > div > div')
-    .forEach(decorateBlock);
-}
-
-/**
- * Builds a block DOM Element from a two dimensional array, string, or object
- * @param {string} blockName name of the block
- * @param {*} content two dimensional array or string or object of content
- */
-export function buildBlock(blockName, content) {
-  const table = Array.isArray(content) ? content : [[content]];
-  const blockEl = document.createElement('div');
-  // build image block nested div structure
-  blockEl.classList.add(blockName);
-  table.forEach((row) => {
-    const rowEl = document.createElement('div');
-    row.forEach((col) => {
-      const colEl = document.createElement('div');
-      const vals = col.elems ? col.elems : [col];
-      vals.forEach((val) => {
-        if (val) {
-          if (typeof val === 'string') {
-            colEl.innerHTML += val;
-          } else {
-            colEl.appendChild(val);
-          }
-        }
-      });
-      rowEl.appendChild(colEl);
-    });
-    blockEl.appendChild(rowEl);
-  });
-  return (blockEl);
-}
-
-/**
- * Loads JS and CSS for a module and executes it's default export.
- * @param {string} name The module name
- * @param {string} jsPath The JS file to load
- * @param {string} [cssPath] An optional CSS file to load
- * @param {object[]} [args] Parameters to be passed to the default export when it is called
- */
-async function loadModule(name, jsPath, cssPath, ...args) {
-  const cssLoaded = cssPath ? loadCSS(cssPath) : Promise.resolve();
-  const decorationComplete = jsPath
-    ? new Promise((resolve) => {
-      (async () => {
-        let mod;
-        try {
-          mod = await import(jsPath);
-          if (mod.default) {
-            await mod.default.apply(null, args);
-          }
-        } catch (error) {
-          // eslint-disable-next-line no-console
-          console.log(`failed to load module for ${name}`, error);
-        }
-        resolve(mod);
-      })();
-    })
-    : Promise.resolve();
-  return Promise.all([cssLoaded, decorationComplete])
-    .then(([, api]) => api);
-}
-
-/**
- * Gets the configuration for the given block, and also passes
- * the config through all custom patching helpers added to the project.
- *
- * @param {Element} block The block element
- * @returns {Object} The block config (blockName, cssPath and jsPath)
- */
-function getBlockConfig(block) {
-  const { blockName } = block.dataset;
-  const jsPath = `${window.hlx.codeBasePath}/blocks/${blockName}/${blockName}.js`;
-  const original = { blockName, jsPath };
-  return (window.hlx.patchBlockConfig || [])
-    .filter((fn) => typeof fn === 'function')
-    .reduce((config, fn) => fn(config, original), { blockName, jsPath });
-}
-
-/**
- * Loads JS and CSS for a block.
- * @param {Element} block The block element
- */
-export async function loadBlock(block) {
-  const status = block.dataset.blockStatus;
-  if (status !== 'loading' && status !== 'loaded') {
-    block.dataset.blockStatus = 'loading';
-    const { blockName, jsPath } = getBlockConfig(block);
-    try {
-      await loadModule(blockName, jsPath, undefined, block);
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      console.log(`failed to load block ${blockName}`, error);
-    }
-    block.dataset.blockStatus = 'loaded';
-  }
-}
-
-/**
- * Loads JS and CSS for all blocks in a container element.
- * @param {Element} main The container element
- */
-export async function loadBlocks(main) {
-  updateSectionsStatus(main);
-  const blocks = [...main.querySelectorAll('div.block')];
-  for (let i = 0; i < blocks.length; i += 1) {
-    // eslint-disable-next-line no-await-in-loop
-    await loadBlock(blocks[i]);
-    updateSectionsStatus(main);
-  }
-}
-
-/**
- * Returns a picture element with webp and fallbacks
- * @param {string} src The image URL
- * @param {string} [alt] The image alternative text
- * @param {boolean} [eager] Set loading attribute to eager
- * @param {Array} [breakpoints] Breakpoints and corresponding params (eg. width)
- * @returns {Element} The picture element
- */
-export function createOptimizedPicture(src, alt = '', eager = false, breakpoints = [{ media: '(min-width: 600px)', width: '2000' }, { width: '750' }]) {
-  const url = new URL(src, window.location.href);
-  const picture = document.createElement('picture');
-  const { pathname } = url;
-  const ext = pathname.substring(pathname.lastIndexOf('.') + 1);
-
-  // webp
-  breakpoints.forEach((br) => {
-    const source = document.createElement('source');
-    if (br.media) source.setAttribute('media', br.media);
-    source.setAttribute('type', 'image/webp');
-    source.setAttribute('srcset', `${pathname}?width=${br.width}&format=webply&optimize=medium`);
-    picture.appendChild(source);
-  });
-
-  // fallback
-  breakpoints.forEach((br, i) => {
-    if (i < breakpoints.length - 1) {
-      const source = document.createElement('source');
-      if (br.media) source.setAttribute('media', br.media);
-      source.setAttribute('srcset', `${pathname}?width=${br.width}&format=${ext}&optimize=medium`);
-      picture.appendChild(source);
-    } else {
-      const img = document.createElement('img');
-      img.setAttribute('loading', eager ? 'eager' : 'lazy');
-      img.setAttribute('alt', alt);
-      picture.appendChild(img);
-      img.setAttribute('src', `${pathname}?width=${br.width}&format=${ext}&optimize=medium`);
-    }
-  });
-
-  return picture;
-}
-
-/**
- * Normalizes all headings within a container element.
- * @param {Element} el The container element
- * @param {string} allowedHeadings The list of allowed headings (h1 ... h6)
- */
-export function normalizeHeadings(el, allowedHeadings) {
-  const allowed = allowedHeadings.map((h) => h.toLowerCase());
-  el.querySelectorAll('h1, h2, h3, h4, h5, h6').forEach((tag) => {
-    const h = tag.tagName.toLowerCase();
-    if (allowed.indexOf(h) === -1) {
-      // current heading is not in the allowed list -> try first to "promote" the heading
-      let level = parseInt(h.charAt(1), 10) - 1;
-      while (allowed.indexOf(`h${level}`) === -1 && level > 0) {
-        level -= 1;
-      }
-      if (level === 0) {
-        // did not find a match -> try to "downgrade" the heading
-        while (allowed.indexOf(`h${level}`) === -1 && level < 7) {
-          level += 1;
-        }
-      }
-      if (level !== 7) {
-        tag.outerHTML = `<h${level} id="${tag.id}">${tag.textContent}</h${level}>`;
-      }
-    }
-  });
-}
-
-/**
- * Set template (page structure) and theme (page styles).
- */
-export function decorateTemplateAndTheme() {
-  const addClasses = (element, classes) => {
-    classes.split(',').forEach((c) => {
-      element.classList.add(toClassName(c.trim()));
-    });
-  };
-  const template = getMetadata('template');
-  if (template) addClasses(document.body, template);
-  const theme = getMetadata('theme');
-  if (theme) addClasses(document.body, theme);
-}
-
-/**
- * Decorates paragraphs containing a single link as buttons.
- * @param {Element} element container element
- */
-export function decorateButtons(element) {
-  element.querySelectorAll('a').forEach((a) => {
-    a.title = a.title || a.textContent;
-    if (a.href !== a.textContent && a.title !== 'link') {
-      const up = a.parentElement;
-      const twoup = a.parentElement.parentElement;
-      if (!a.querySelector('img') && twoup.tagName !== 'LI' && !a.closest('.call-to-action')
-          && !a.closest('.mini-teasers') && !a.closest('.bg-color-right')) {
-        if (up.childNodes.length === 1 && (up.tagName === 'P' || up.tagName === 'DIV')) {
-          a.className = 'btn btn-outline-primary'; // default
-          up.classList.add('button-container');
-        }
-        if (up.childNodes.length === 1 && up.tagName === 'STRONG'
-          && twoup.childNodes.length === 1 && twoup.tagName === 'P') {
-          a.className = 'btn btn-outline-primary';
-          twoup.classList.add('button-container');
-        }
-        if (up.childNodes.length === 1 && up.tagName === 'EM'
-          && twoup.childNodes.length === 1 && twoup.tagName === 'P') {
-          a.className = 'btn btn-outline-secondary';
-          twoup.classList.add('button-container');
-        }
-      }
-    }
-  });
-}
-
-/**
- * Load LCP block and/or wait for LCP in default content.
- */
-export async function waitForLCP(lcpBlocks) {
-  const block = document.querySelector('.block');
-  const hasLCPBlock = (block && lcpBlocks.includes(block.dataset.blockName));
-  if (hasLCPBlock) await loadBlock(block);
-
-  document.body.style.display = null;
-  const lcpCandidate = document.querySelector('main img');
-  await new Promise((resolve) => {
-    if (lcpCandidate && !lcpCandidate.complete) {
-      lcpCandidate.setAttribute('loading', 'eager');
-      lcpCandidate.addEventListener('load', resolve);
-      lcpCandidate.addEventListener('error', resolve);
-    } else {
-      resolve();
-    }
-  });
-}
-
-/**
- * Loads a block named 'header' into header
- * @param {Element} header header element
- * @returns {Promise}
- */
-export function loadHeader(header) {
-  const headerBlock = buildBlock('header', '');
-  header.append(headerBlock);
-  decorateBlock(headerBlock);
-  return loadBlock(headerBlock);
-}
-
-/**
- * Loads a block named 'footer' into footer
- * @param footer footer element
- * @returns {Promise}
- */
-export function loadFooter(footer) {
-  const footerBlock = buildBlock('footer', '');
-  footer.append(footerBlock);
-  decorateBlock(footerBlock);
-  return loadBlock(footerBlock);
-}
-
-function parsePluginParams(id, config) {
-  const pluginId = !config
-    ? id.split('/').splice(id.endsWith('/') ? -2 : -1, 1)[0].replace(/\.js/, '')
-    : id;
-  const pluginConfig = {
-    load: 'eager',
-    ...(typeof config === 'string' || !config
-      ? { url: (config || id).replace(/\/$/, '') }
-      : config),
-  };
-  pluginConfig.options ||= {};
-  return { id: pluginId, config: pluginConfig };
-}
-
-// Define an execution context for plugins
-export const executionContext = {
-  createOptimizedPicture,
-  getAllMetadata,
-  getMetadata,
-  decorateBlock,
-  decorateButtons,
-  decorateIcons,
-  loadBlock,
-  loadCSS,
-  loadScript,
-  sampleRUM,
-  toCamelCase,
-  toClassName,
-};
-
-class PluginsRegistry {
-  #plugins;
-
-  constructor() {
-    this.#plugins = new Map();
-  }
-
-  // Register a new plugin
-  add(id, config) {
-    const { id: pluginId, config: pluginConfig } = parsePluginParams(id, config);
-    this.#plugins.set(pluginId, pluginConfig);
-  }
-
-  // Get the plugin
-  get(id) { return this.#plugins.get(id); }
-
-  // Check if the plugin exists
-  includes(id) { return !!this.#plugins.has(id); }
-
-  // Load all plugins that are referenced by URL, and updated their configuration with the
-  // actual API they expose
-  async load(phase) {
-    [...this.#plugins.entries()]
-      .filter(([, plugin]) => plugin.condition
-      && !plugin.condition(document, plugin.options, executionContext))
-      .map(([id]) => this.#plugins.delete(id));
-    return Promise.all([...this.#plugins.entries()]
-      // Filter plugins that don't match the execution conditions
-      .filter(([, plugin]) => (
-        (!plugin.condition || plugin.condition(document, plugin.options, executionContext))
-        && phase === plugin.load && plugin.url
-      ))
-      .map(async ([key, plugin]) => {
-        try {
-          // If the plugin has a default export, it will be executed immediately
-          const pluginApi = (await loadModule(
-            key,
-            !plugin.url.endsWith('.js') ? `${plugin.url}/${key}.js` : plugin.url,
-            !plugin.url.endsWith('.js') ? `${plugin.url}/${key}.css` : null,
-            document,
-            plugin.options,
-            executionContext,
-          )) || {};
-          this.#plugins.set(key, { ...plugin, ...pluginApi });
-        } catch (err) {
-          // eslint-disable-next-line no-console
-          console.error('Could not load specified plugin', key);
-        }
-      }));
-  }
-
-  // Run a specific phase in the plugin
-  async run(phase) {
-    return [...this.#plugins.values()]
-      .reduce((promise, plugin) => ( // Using reduce to execute plugins sequencially
-        plugin[phase] && (!plugin.condition
-            || plugin.condition(document, plugin.options, executionContext))
-          ? promise.then(() => plugin[phase](document, plugin.options, executionContext))
-          : promise
-      ), Promise.resolve());
-  }
-}
-
-class TemplatesRegistry {
-  // Register a new template
-  // eslint-disable-next-line class-methods-use-this
-  add(id, url) {
-    if (Array.isArray(id)) {
-      id.forEach((i) => window.hlx.templates.add(i));
-      return;
-    }
-    const { id: templateId, config: templateConfig } = parsePluginParams(id, url);
-    templateConfig.condition = () => toClassName(getMetadata('template')) === templateId;
-    window.hlx.plugins.add(templateId, templateConfig);
-  }
-
-  // Get the template
-  // eslint-disable-next-line class-methods-use-this
-  get(id) { return window.hlx.plugins.get(id); }
-
-  // Check if the template exists
-  // eslint-disable-next-line class-methods-use-this
-  includes(id) { return window.hlx.plugins.includes(id); }
-}
-
-/**
- * Setup block utils.
- */
-export function setup() {
-  window.hlx = window.hlx || {};
-  window.hlx.RUM_MASK_URL = 'full';
-  window.hlx.codeBasePath = '';
-  window.hlx.lighthouse = new URLSearchParams(window.location.search).get('lighthouse') === 'on';
-  window.hlx.patchBlockConfig = [];
-  window.hlx.plugins = new PluginsRegistry();
-  window.hlx.templates = new TemplatesRegistry();
-
-  const scriptEl = document.querySelector('script[src$="/scripts/scripts.js"]');
-  if (scriptEl) {
-    try {
-      [window.hlx.codeBasePath] = new URL(scriptEl.src).pathname.split('/scripts/scripts.js');
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      console.log(error);
-    }
-  }
-}
-
-/**
- * Auto initializiation.
- */
-function init() {
-  setup();
-  sampleRUM('top');
-
-  window.addEventListener('load', () => sampleRUM('load'));
-
-  window.addEventListener('unhandledrejection', (event) => {
-    sampleRUM('error', { source: event.reason.sourceURL, target: event.reason.line });
-  });
-
-  window.addEventListener('error', (event) => {
-    sampleRUM('error', { source: event.filename, target: event.lineno });
-  });
-}
-
-init();
+function sampleRUM(e,t={}){sampleRUM.defer=sampleRUM.defer||[];var a,o,n,l,r,s,i,c,d,u=t=>{sampleRUM[t]=sampleRUM[t]||((...e)=>sampleRUM.defer.push({fnname:t,args:e}))};sampleRUM.drain=sampleRUM.drain||((t,e)=>{sampleRUM[t]=e,sampleRUM.defer.filter(({fnname:e})=>t===e).forEach(({fnname:e,args:t})=>sampleRUM[e](...t))}),sampleRUM.always=sampleRUM.always||[],sampleRUM.always.on=(e,t)=>{sampleRUM.always[e]=t},sampleRUM.on=(e,t)=>{sampleRUM.cases[e]=t},u("observe"),u("cwv");try{window.hlx=window.hlx||{},window.hlx.rum||(a="on"===new URLSearchParams(window.location.search).get("rum")?1:20,o=Array.from({length:75},(e,t)=>String.fromCharCode(48+t)).filter(e=>/\d|[A-Z]/i.test(e)).filter(()=>70<75*Math.random()).join(""),l=(n=Math.random())*a<1,r=Date.now(),s={full:()=>window.location.href,origin:()=>window.location.origin,path:()=>window.location.href.replace(/\?.*$/,"")},window.hlx.rum={weight:a,id:o,random:n,isSelected:l,firstReadTime:r,sampleRUM:sampleRUM,sanitizeURL:s[window.hlx.RUM_MASK_URL||"path"]});const{weight:m,id:p,firstReadTime:h}=window.hlx.rum;if(window.hlx&&window.hlx.rum&&window.hlx.rum.isSelected){const w=["weight","id","referer","checkpoint","t","source","target","cwv","CLS","FID","LCP","INP"];sampleRUM.cases=sampleRUM.cases||{cwv:()=>sampleRUM.cwv(t)||!0,lazy:()=>{var e=document.createElement("script");return e.src="https://rum.hlx.page/.rum/@adobe/helix-rum-enhancer@^1/src/index.js",document.head.appendChild(e),!0}},[i=t]=[t],c=JSON.stringify({weight:m,id:p,referer:window.hlx.rum.sanitizeURL(),checkpoint:e,t:Date.now()-h,...t},w),d="https://rum.hlx.page/.rum/"+m,navigator.sendBeacon(d,c),console.debug("ping:"+e,i),sampleRUM.cases[e]&&sampleRUM.cases[e]()}sampleRUM.always[e]&&sampleRUM.always[e](t)}catch(e){}}async function loadCSS(o){return new Promise((e,t)=>{var a;document.querySelector(`head > link[href="${o}"]`)?e():((a=document.createElement("link")).rel="stylesheet",a.href=o,a.onload=e,a.onerror=t,document.head.append(a))})}async function loadScript(n,l){return new Promise((e,t)=>{if(document.querySelector(`head > script[src="${n}"]`))e();else{var a=document.createElement("script");if(a.src=n,l)for(const o in l)a.setAttribute(o,l[o]);a.onload=e,a.onerror=t,document.head.append(a)}})}function getMetadata(e){var t=e&&e.includes(":")?"property":"name";return[...document.head.querySelectorAll(`meta[${t}="${e}"]`)].map(e=>e.content).join(", ")||""}function toClassName(e){return"string"==typeof e?e.toLowerCase().replace(/[^0-9a-z]/gi,"-").replace(/-+/g,"-").replace(/^-|-$/g,""):""}function toCamelCase(e){return toClassName(e).replace(/-([a-z])/g,e=>e[1].toUpperCase())}function getAllMetadata(a){return[...document.head.querySelectorAll(`meta[property^="${a}:"],meta[name^="${a}-"]`)].reduce((e,t)=>{return e[toClassName(t.name?t.name.substring(a.length+1):t.getAttribute("property").split(":")[1])]=t.getAttribute("content"),e},{})}const ICONS_CACHE={};async function decorateIcons(e){let t=document.getElementById("franklin-svg-sprite");t||((a=document.createElement("div")).innerHTML='<svg xmlns="http://www.w3.org/2000/svg" id="franklin-svg-sprite" style="display: none"></svg>',t=a.firstElementChild,document.body.append(a.firstElementChild));var a=[...e.querySelectorAll("span.icon")],e=(await Promise.all(a.map(async t=>{t=Array.from(t.classList).find(e=>e.startsWith("icon-")).substring(5);if(!ICONS_CACHE[t]){ICONS_CACHE[t]=!0;try{let e=window.hlx.codeBasePath+`/icons/${t}.svg`;t.startsWith("dam-")&&(a=window.location.hostname.includes("lifesciences.danaher.com"),e=a?"":"https://lifesciences.danaher.com",e+=`/content/dam/danaher/system/icons/${t.substring(4).replace("_"," ")}.svg`);var a,o,n=await fetch(e);n.ok?(o=await n.text()).match(/(<style | class=)/)?ICONS_CACHE[t]={styled:!0,html:o}:ICONS_CACHE[t]={html:o.replace("<svg",`<symbol id="icons-sprite-${t}"`).replace(/ width=".*?"/,"").replace(/ height=".*?"/,"").replace("</svg>","</symbol>")}:ICONS_CACHE[t]=!1}catch(e){ICONS_CACHE[t]=!1,console.error(e)}}})),Object.keys(ICONS_CACHE).filter(e=>!t.querySelector("#icons-sprite-"+e)).map(e=>ICONS_CACHE[e]).filter(e=>!e.styled).map(e=>e.html).join("\n"));t.innerHTML+=e,a.forEach(e=>{var t=Array.from(e.classList).find(e=>e.startsWith("icon-")).substring(5),e="A"===e.firstElementChild?.tagName?e.firstElementChild:e;ICONS_CACHE[t].styled?e.innerHTML=ICONS_CACHE[t].html:e.innerHTML=`<svg xmlns="http://www.w3.org/2000/svg"><use href="#icons-sprite-${t}"/></svg>`})}async function fetchPlaceholders(o="default"){return window.placeholders=window.placeholders||{},window.placeholders[o+"-loaded"]||(window.placeholders[o+"-loaded"]=new Promise((a,t)=>{fetch(`${"default"===o?"":o}/placeholders.json`).then(e=>{if(e.ok)return e.json();throw new Error(e.status+": "+e.statusText)}).then(e=>{const t={};e.data.filter(e=>e.Key).forEach(e=>{t[toCamelCase(e.Key)]=e.Text}),window.placeholders[o]=t,a()}).catch(e=>{window.placeholders[o]={},t(e)})})),await window.placeholders[o+"-loaded"],window.placeholders[o]}function decorateBlock(e){var t=e.classList[0];t&&(e.classList.add("block"),e.dataset.blockName=t,e.dataset.blockStatus="initialized",e.parentElement.classList.add(t+"-wrapper"),e=e.closest(".section"))&&e.classList.add(t+"-container")}function readBlockConfig(e){const l={};return e.querySelectorAll(":scope > div").forEach(t=>{if(t.children){var a=[...t.children];if(a[1]){var o,n=a[1],a=toClassName(a[0].textContent);let e="";e=n.querySelector("a")?1===(o=[...n.querySelectorAll("a")]).length?o[0].href:o.map(e=>e.href):n.querySelector("img")?1===(o=[...n.querySelectorAll("img")]).length?o[0].src:o.map(e=>e.src):n.querySelector("p")?1===(o=[...n.querySelectorAll("p")]).length?o[0].textContent:o.map(e=>e.textContent):t.children[1].textContent,l[a]=e}}}),l}function decorateSections(e){e.querySelectorAll(":scope > div").forEach(t=>{const a=[];let o=!1;[...t.children].forEach(e=>{var t;"DIV"!==e.tagName&&o||(t=document.createElement("div"),a.push(t),(o="DIV"!==e.tagName)&&t.classList.add("default-content-wrapper")),a[a.length-1].append(e)}),a.forEach(e=>t.append(e)),t.classList.add("section"),t.dataset.sectionStatus="initialized",t.style.display="none";var e=t.querySelector("div.section-metadata");if(e){const n=readBlockConfig(e);Object.keys(n).forEach(e=>{"style"===e?n.style.split(",").map(e=>toClassName(e.trim())).forEach(e=>t.classList.add(e)):t.dataset[toCamelCase(e)]=n[e]}),e.parentNode.remove()}})}function updateSectionsStatus(e){var t=[...e.querySelectorAll(":scope > div.section")];for(let e=0;e<t.length;e+=1){var a=t[e];if("loaded"!==a.dataset.sectionStatus){if(a.querySelector('.block[data-block-status="initialized"], .block[data-block-status="loading"]')){a.dataset.sectionStatus="loading";break}a.dataset.sectionStatus="loaded",a.style.display=null}}}function decorateBlocks(e){e.querySelectorAll("div.section > div > div").forEach(decorateBlock)}function buildBlock(e,t){t=Array.isArray(t)?t:[[t]];const o=document.createElement("div");return o.classList.add(e),t.forEach(e=>{const a=document.createElement("div");e.forEach(e=>{const t=document.createElement("div");(e.elems||[e]).forEach(e=>{e&&("string"==typeof e?t.innerHTML+=e:t.appendChild(e))}),a.appendChild(t)}),o.appendChild(a)}),o}async function loadModule(a,o,e,...n){var e=e?loadCSS(e):Promise.resolve(),t=o?new Promise(t=>{(async()=>{let e;try{(e=await import(o)).default&&await e.default.apply(null,n)}catch(e){console.log("failed to load module for "+a,e)}t(e)})()}):Promise.resolve();return Promise.all([e,t]).then(([,e])=>e)}function getBlockConfig(e){var e=e.dataset["blockName"],t=window.hlx.codeBasePath+`/blocks/${e}/${e}.js`;const a={blockName:e,jsPath:t};return(window.hlx.patchBlockConfig||[]).filter(e=>"function"==typeof e).reduce((e,t)=>t(e,a),{blockName:e,jsPath:t})}async function loadBlock(e){var t=e.dataset.blockStatus;if("loading"!==t&&"loaded"!==t){e.dataset.blockStatus="loading";var{blockName:t,jsPath:a}=getBlockConfig(e);try{await loadModule(t,a,void 0,e)}catch(e){console.log("failed to load block "+t,e)}e.dataset.blockStatus="loaded"}}async function loadBlocks(t){updateSectionsStatus(t);var a=[...t.querySelectorAll("div.block")];for(let e=0;e<a.length;e+=1)await loadBlock(a[e]),updateSectionsStatus(t)}function createOptimizedPicture(e,a="",o=!1,n=[{media:"(min-width: 600px)",width:"2000"},{width:"750"}]){e=new URL(e,window.location.href);const l=document.createElement("picture"),r=e["pathname"],s=r.substring(r.lastIndexOf(".")+1);return n.forEach(e=>{var t=document.createElement("source");e.media&&t.setAttribute("media",e.media),t.setAttribute("type","image/webp"),t.setAttribute("srcset",`${r}?width=${e.width}&format=webply&optimize=medium`),l.appendChild(t)}),n.forEach((e,t)=>{t<n.length-1?(t=document.createElement("source"),e.media&&t.setAttribute("media",e.media),t.setAttribute("srcset",`${r}?width=${e.width}&format=${s}&optimize=medium`),l.appendChild(t)):((t=document.createElement("img")).setAttribute("loading",o?"eager":"lazy"),t.setAttribute("alt",a),l.appendChild(t),t.setAttribute("src",`${r}?width=${e.width}&format=${s}&optimize=medium`))}),l}function normalizeHeadings(e,t){const o=t.map(e=>e.toLowerCase());e.querySelectorAll("h1, h2, h3, h4, h5, h6").forEach(t=>{var a=t.tagName.toLowerCase();if(-1===o.indexOf(a)){let e=parseInt(a.charAt(1),10)-1;for(;-1===o.indexOf("h"+e)&&0<e;)--e;if(0===e)for(;-1===o.indexOf("h"+e)&&e<7;)e+=1;7!==e&&(t.outerHTML=`<h${e} id="${t.id}">${t.textContent}</h${e}>`)}})}function decorateTemplateAndTheme(){var e=(t,e)=>{e.split(",").forEach(e=>{t.classList.add(toClassName(e.trim()))})},t=getMetadata("template"),t=(t&&e(document.body,t),getMetadata("theme"));t&&e(document.body,t)}function decorateButtons(e){e.querySelectorAll("a").forEach(e=>{var t,a;e.title=e.title||e.textContent,e.href!==e.textContent&&"link"!==e.title&&(t=e.parentElement,a=e.parentElement.parentElement,e.querySelector("img")||"LI"===a.tagName||e.closest(".call-to-action")||e.closest(".mini-teasers")||e.closest(".bg-color-right")||(1!==t.childNodes.length||"P"!==t.tagName&&"DIV"!==t.tagName||(e.className="btn btn-outline-primary",t.classList.add("button-container")),1===t.childNodes.length&&"STRONG"===t.tagName&&1===a.childNodes.length&&"P"===a.tagName&&(e.className="btn btn-outline-primary",a.classList.add("button-container")),1===t.childNodes.length&&"EM"===t.tagName&&1===a.childNodes.length&&"P"===a.tagName&&(e.className="btn btn-outline-secondary",a.classList.add("button-container"))))})}async function waitForLCP(e){var t=document.querySelector(".block");t&&e.includes(t.dataset.blockName)&&await loadBlock(t),document.body.style.display=null;const a=document.querySelector("main img");await new Promise(e=>{a&&!a.complete?(a.setAttribute("loading","eager"),a.addEventListener("load",e),a.addEventListener("error",e)):e()})}function loadHeader(e){var t=buildBlock("header","");return e.append(t),decorateBlock(t),loadBlock(t)}function loadFooter(e){var t=buildBlock("footer","");return e.append(t),decorateBlock(t),loadBlock(t)}function parsePluginParams(e,t){var a=t?e:e.split("/").splice(e.endsWith("/")?-2:-1,1)[0].replace(/\.js/,""),t={load:"eager",..."string"!=typeof t&&t?t:{url:(t||e).replace(/\/$/,"")}};return t.options||={},{id:a,config:t}}const executionContext={createOptimizedPicture:createOptimizedPicture,getAllMetadata:getAllMetadata,getMetadata:getMetadata,decorateBlock:decorateBlock,decorateButtons:decorateButtons,decorateIcons:decorateIcons,loadBlock:loadBlock,loadCSS:loadCSS,loadScript:loadScript,sampleRUM:sampleRUM,toCamelCase:toCamelCase,toClassName:toClassName};class PluginsRegistry{#plugins;constructor(){this.#plugins=new Map}add(e,t){var{id:e,config:t}=parsePluginParams(e,t);this.#plugins.set(e,t)}get(e){return this.#plugins.get(e)}includes(e){return!!this.#plugins.has(e)}async load(t){return[...this.#plugins.entries()].filter(([,e])=>e.condition&&!e.condition(document,e.options,executionContext)).map(([e])=>this.#plugins.delete(e)),Promise.all([...this.#plugins.entries()].filter(([,e])=>(!e.condition||e.condition(document,e.options,executionContext))&&t===e.load&&e.url).map(async([t,e])=>{try{var a=await loadModule(t,e.url.endsWith(".js")?e.url:e.url+`/${t}.js`,e.url.endsWith(".js")?null:e.url+`/${t}.css`,document,e.options,executionContext)||{};this.#plugins.set(t,{...e,...a})}catch(e){console.error("Could not load specified plugin",t)}}))}async run(a){return[...this.#plugins.values()].reduce((e,t)=>!t[a]||t.condition&&!t.condition(document,t.options,executionContext)?e:e.then(()=>t[a](document,t.options,executionContext)),Promise.resolve())}}class TemplatesRegistry{add(e,t){if(Array.isArray(e))e.forEach(e=>window.hlx.templates.add(e));else{const{id:a,config:o}=parsePluginParams(e,t);o.condition=()=>toClassName(getMetadata("template"))===a,window.hlx.plugins.add(a,o)}}get(e){return window.hlx.plugins.get(e)}includes(e){return window.hlx.plugins.includes(e)}}function setup(){window.hlx=window.hlx||{},window.hlx.RUM_MASK_URL="full",window.hlx.codeBasePath="",window.hlx.lighthouse="on"===new URLSearchParams(window.location.search).get("lighthouse"),window.hlx.patchBlockConfig=[],window.hlx.plugins=new PluginsRegistry,window.hlx.templates=new TemplatesRegistry;var e=document.querySelector('script[src$="/scripts/scripts.js"]');if(e)try{[window.hlx.codeBasePath]=new URL(e.src).pathname.split("/scripts/scripts.js")}catch(e){console.log(e)}}function init(){setup(),sampleRUM("top"),window.addEventListener("load",()=>sampleRUM("load")),window.addEventListener("unhandledrejection",e=>{sampleRUM("error",{source:e.reason.sourceURL,target:e.reason.line})}),window.addEventListener("error",e=>{sampleRUM("error",{source:e.filename,target:e.lineno})})}init();export{sampleRUM,loadCSS,loadScript,getMetadata,toClassName,toCamelCase,getAllMetadata,decorateIcons,fetchPlaceholders,decorateBlock,readBlockConfig,decorateSections,updateSectionsStatus,decorateBlocks,buildBlock,loadBlock,loadBlocks,createOptimizedPicture,normalizeHeadings,decorateTemplateAndTheme,decorateButtons,waitForLCP,loadHeader,loadFooter,executionContext,setup};
